fix: address code review findings for Phase 4 JS/WASM
- bootstrap.go: release js.FuncOf handler after Promise construction to
  prevent monotonic leak of Go function table slots on each call
- params.go: return float64 from ckksGaloisElement for consistency with
  ckksModuliChain/ckksAuxModuliChain (uint64 → int64 was inconsistent)
- server/main.go: fix HandleInfer data race — serialize Forward() calls
  per session with per-session mutex; use crypto/rand for session IDs
  instead of sequential integers; add http.MaxBytesReader caps (512 MB
  for key sets, 32 MB for ciphertexts); cache ckks.Parameters in Server
  at construction time; NewServer now returns an error
- server/main_test.go: fix ignored json.Decode error returns; use
  distinct input values in TestFullRoundtrip; add TestCORSMiddleware;
  update testSetup for new NewServer/CKKSParams signatures
- tests/helpers.ts: replace boolean flag with promise singleton to
  prevent double-initialization race in concurrent beforeAll hooks
- client.ts: clarify that the bootstrap block is dead code in this demo
- ARCH.md: document JS/WASM build commands, JS test command, and
  browser demo run instructions
- Move completed plan to docs/plans/completed/
- Add README for js/lattigo and examples/wasm-demo

+++ js/lattigo/bridge/params.go
@@ -111,7 +111,9 @@ func ckksGaloisElement(_ js.Value, args []js.Value) any {
 		return nil
 	}
 	rotation := args[1].Int()
-	return int64(p.(*ckks.Parameters).GaloisElement(rotation))
+	// Return float64 for consistency with ckksModuliChain/ckksAuxModuliChain.
+	// Galois elements fit safely in float64 for any practical CKKS ring dimension.
+	return float64(p.(*ckks.Parameters).GaloisElement(rotation))
 }
 
 // ckksModuliChain(paramsHID: number) → Array<number>
